[core] Make StateMachineCache public and hide AsyncVoidMethodBuilder
Firstly, make the built-in AsyncVoidMethodBuilder implementation
internal, not public. I didn't realise the compiler restriction
was that alternative AsyncVoidMethodBuilder have to be declared
within the source code of the project for them to be chosen in
preference to the normal built-in one.

If users of ResuableTasks want to use an implementation of
AsyncVoidMethodBuilder which is based on ReusableTask (for the memory
allocation benefits) then they'll need to be able to access
StateMachineCache.