[core] Add some sanity checks to cover bad usage of ReusableTask
There are certain invariants I think I can rely on :p

Normal usage of async methods ensures that several methods are
only invoked once, if people obey the restrictions enforced by
ReusableTask. Now that the various classes are value types, it
is trivial to add a token ensuring that these methods are not
invoked more than once.

While this is not foolproof (as value types are copied by value),
it may be sufficient to catch the majority of cases.