[core] Use Interlocked.CompareExchange to update ResultHolder
This should be a tad cheaper than using a full lock. It's
incredibly unlikely this would ever be contended though, as
either the ResultHolder will have it's value set synchronously,
or the continuation object will be set before the asynchronous
value is set. The odds of the async value being generated
within a few clock cycles are pretty slim :p