[core] Improve performance for synchronously completed tasks
Tasks which complete synchronously will no longer use a
ResultHolder object. This is a performance benefit in two
ways. Firstly we do not have to retrieve an object from
the static cache, nor will we need to use any Interlocked
operations to set the value and/or continuation.

The continuation is also guaranteed to be executed
synchronously too, just like it would in a normal
ValueTask or Task which has completed synchronously.