[core] Make the main structs readonly, and adjust accordingly
If structs are readonly then it allows the compiler to avoid
creating defensive copies.

The first stage in this change was to mark all the structs,
except the Builders, as readonly. To make it compile the
'Id' field had to be made readonly too. This meant I had to
rethink the way reuse was verified.

The readonly nature of all the structs makes it easier to reason
about. A ReusableTask is always instantiated with a ResultHolder.
That ResultHolder is a class with an Id which changes every time
the ResultHolder is reset. If the Id is captured when constructing
the ReusableTask, then we can store it and validate it when
interacting with the ResultHolder later on.

This means we have a better/stronger check against accidental
bad usage of ReusableTask than previously.

ResultHolder.Id is implemented using 30 bits from the state
variable, which means we don't pay any additional memory
penalty for this feature.