[core] Simplify how completion occurs
The main aim of this commit is to ensure that 'HasValue',
'TrySetExceptionOrResult' and the setter for 'Continuation'
all have an *identical* semantic for when a task is complete
or not.

The idea is this: We know 'IsCompleted' is invoked once, to
determine whether the task has completed syncronously. If
it has completed synchronously then GetResult will be
invoked immediately and the ReusableTask will be recycled.
Otherwise 'OnCompleted' will be invoked and we'll receive a
delegate to invoke at a future point in time.

The issue with the original code was that it is possible
for 'TrySetResult' to execute at the same time as
'IsCompleted' is being checked. If the thread setting the
result is suspended right after it sets the 'HasValue' flag
and right before it sets the 'Continuation' to the
'HasValueSentinel' then it is plausible for the task to
complete synchronously and be reset so it can be reused,
before the original thread resumes execution and sets the
Continuation field to 'HasValueSentinel'.

If this occurs then we'll end up invoking the HasValueSentinel
the next time SetResult is invoked, which throws an Exception.

This changes the logic so that IsCompleted only returns true
if the Continuation has been set to 'HasValueSentinel'. This
means that the only time it is true is if SetResult has been
invoked *before* 'IsCompleted' is checked, which means we'll
complete synchronously.

If we are in an async completion scenario, then IsCompleted
will always return false as the 'Continuation' property
will never be equal to 'HasValueSentinel' again, but that's
OK because we'll immediately invoke the continuation as soon
as SetResult or SetException is called and 'IsCompleted' will
not be checked again.