[core] Remove the public TrySet methods
These could tempt people into using the library badly.
You must call one of these exactly once for every 'await'
call. Typically TrySet methods are used to cancel pending
requests, but if that is used on a ReusableTaskCompletionSource
it will instead cancel the current object, as expected, but then
the actual value will be set later which is *unexpected*.