Use spinlocks in low-contention areas
There are some places where a lock is needed just to push or pop
something from a stack. In these cases, a spinlock appears to be more
performant than a regular lock. It's a hard one to benchmark :p

This performs worse in cases where you have a tight loop
starting/completing tasks, which is exactly what you'd do in a
benchmark. Though in most real world situations there'll be plenty of
code in the async method so this is unlikely to be the actual point of
contention.